#341 Refactor: Navbar as an atom
